# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ALT Agency Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List all audits for current workspace
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audit"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Start a new website audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditInput"
      responses:
        "201":
          description: Audit started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
  /audits/{id}:
    get:
      operationId: getAudit
      tags: [audits]
      summary: Get audit by ID with full results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditDetail"
    delete:
      operationId: deleteAudit
      tags: [audits]
      summary: Delete an audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /audits/{id}/regenerate:
    post:
      operationId: regenerateAuditReport
      tags: [audits]
      summary: Regenerate AI report for audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRegenerateInput"
      responses:
        "200":
          description: Regenerated audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditDetail"
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by ID or slug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Billing ────────────────────────────────────────────────────────────────
...
  /billing/waitlist:
    post:
      operationId: joinWaitlist
      tags: [billing]
      summary: Join the product waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "201":
          description: Joined waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
components:
  schemas:
...
    Audit:
      type: object
      required: [id, workspaceId, url, status, overallScore, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        url:
          type: string
        websiteName:
          type: ["string", "null"]
        status:
          type: string
        overallScore:
          type: ["integer", "null"]
        seoScore:
          type: ["integer", "null"]
        performanceScore:
          type: ["integer", "null"]
        accessibilityScore:
          type: ["integer", "null"]
        uxScore:
          type: ["integer", "null"]
        screenshotUrl:
          type: ["string", "null"]
        issueCount:
          type: integer
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    AuditDetail:
      type: object
      required: [id, workspaceId, url, status, overallScore, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        url:
          type: string
        websiteName:
          type: ["string", "null"]
        status:
          type: string
        overallScore:
          type: ["integer", "null"]
        seoScore:
          type: ["integer", "null"]
        performanceScore:
          type: ["integer", "null"]
        accessibilityScore:
          type: ["integer", "null"]
        uxScore:
          type: ["integer", "null"]
        conversionScore:
          type: ["integer", "null"]
        mobileScore:
          type: ["integer", "null"]
        screenshotUrl:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        aiRecommendations:
          type: ["string", "null"]
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AuditIssue"
        issueCount:
          type: integer
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    AuditIssue:
      type: object
      required: [id, category, severity, title, description]
      properties:
        id:
          type: integer
        auditId:
          type: integer
        category:
          type: string
        severity:
          type: string
        title:
          type: string
        description:
          type: string
        recommendation:
          type: ["string", "null"]
        impact:
          type: ["string", "null"]
    AuditInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        tone:
          type: string
        reportStyle:
          type: string
    ReportRegenerateInput:
      type: object
      properties:
        tone:
          type: string
        style:
          type: string
...
    BlogPost:
      type: object
      required: [id, title, slug, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        published:
          type: boolean
        readTime:
          type: ["integer", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        seoTitle:
          type: string
        seoDescription:
          type: string
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        seoTitle:
          type: string
        seoDescription:
          type: string

    # ── Billing ───────────────────────────────────────────────────────────────
...
    WaitlistInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string
        company:
          type: string
    WaitlistEntry:
      type: object
      required: [id, email, position, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        position:
          type: integer
        createdAt:
          type: string