# lib/api-zod/src/generated/types/auditDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ALT Agency Automation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditIssue } from "./auditIssue";

export interface AuditDetail {
  id: number;
  workspaceId: number;
  url: string;
  /** @nullable */
  websiteName?: string | null;
  status: string;
  /** @nullable */
  overallScore: number | null;
  /** @nullable */
  seoScore?: number | null;
  /** @nullable */
  performanceScore?: number | null;
  /** @nullable */
  accessibilityScore?: number | null;
  /** @nullable */
  uxScore?: number | null;
  /** @nullable */
  conversionScore?: number | null;
  /** @nullable */
  mobileScore?: number | null;
  /** @nullable */
  screenshotUrl?: string | null;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  aiRecommendations?: string | null;
  issues?: AuditIssue[];
  issueCount?: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}